refactor(layer:launchdiscovery slice:SLC-09): verify native identity and bounded revalidation

+++ internal/launchdiscovery/stamp_freebsd.go
@@ -7,10 +7,10 @@ import (
 
 func unixStamp(_ int, st *unix.Stat_t, profile string) (string, error) {
 	if profile == "" || strings.HasPrefix(profile, "birth:") {
-		return stamp("birth", st.Btim.Sec, st.Btim.Nsec), nil
+		return stamp("birth", int64(st.Btim.Sec), int64(st.Btim.Nsec)), nil
 	}
 	if strings.HasPrefix(profile, "change:") {
-		return stamp("change", st.Ctim.Sec, st.Ctim.Nsec), nil
+		return stamp("change", int64(st.Ctim.Sec), int64(st.Ctim.Nsec)), nil
 	}
 	return "", errRedirect
 }
